+++ e2e/specs/confirmations/send-erc721.spec.js
@@ -11,9 +11,11 @@ import Accounts from '../../../wdio/helpers/Accounts';
 import Ganache from '../../../app/util/test/ganache';
 import Browser from '../../pages/Drawer/Browser';
 import ConnectModal from '../../pages/modals/ConnectModal';
-import { TEST_DAPP_URL } from '../../pages/TestDApp';
+import { TEST_DAPP_URL, TestDApp } from '../../pages/TestDApp';
 import root from '../../../locales/languages/en.json';
+import messages from '../../../locales/languages/en.json';
 
+const CONFIRM_BUTTON_TEXT = messages.confirmation_modal.confirm_cta;
 const validAccount = Accounts.getValidAccount();
 const ERC721_ADDRESS = '0x26D6C3e7aEFCE970fe3BE5d589DbAbFD30026924';
 
@@ -42,14 +44,17 @@ describe(Regression('Send token tests'), () => {
     await TabBarComponent.tapBrowser();
 
     // Navigate to the ERC721 url
-    await Browser.navigateToErc721Contract(TEST_DAPP_URL, ERC721_ADDRESS);
+    await TestDApp.navigateToErc721Contract(TEST_DAPP_URL, ERC721_ADDRESS);
 
     // Connect account
     await Browser.tapConnectButton();
     await ConnectModal.tapConnectButton();
 
     // Transfer NFT
-    await Browser.tapTransferFromButton(TEST_DAPP_URL, ERC721_ADDRESS);
+    await TestDApp.tapTransferFromButton(ERC721_ADDRESS);
+    await TestHelpers.delay(3000);
+
+    await TestHelpers.tapByText(CONFIRM_BUTTON_TEXT, 0);
 
     // Navigate to the activity screen
     await TabBarComponent.tapActivity();
